Sunsetting `intro.nvim`
To the 1-2 people who ended up using this plugin, Thank you!

Writing this plugin has taught me a lot and I have understood my limits.
I will go learn new things to hopefully make a better version of this
plugin.

This plugin will *NOT* be updated anumore and I will instead be working
on a rewrite that will be more performant and more stable.

Again, thank you for trying this out and goodbye.